Rearranged lobby folder structure and renamed lobby to home for left panel in mobile view
