+++ packages/frontend/components/lobby/leftPanel/home.tsx
@@ -1,8 +1,8 @@
 import { ReactElement } from 'react';
 
-import Taako from '../taako';
-import Announcements from '../announcements';
-import OnlinePlayers from '../onlinePlayers';
+import Taako from './taako';
+import Announcements from './announcements';
+import OnlinePlayers from './onlinePlayers';
 import dateGenObj from '../../../utils/dateGenerator';
 
 const LobbyLeftPanel = (): ReactElement => (
