+++ packages/frontend/components/lobby/lobbyIndex.tsx
@@ -1,8 +1,8 @@
 import React, { ReactElement } from 'react';
 import { connect } from 'react-redux';
 
-import { RootState } from '../../../store';
-import { MobileView } from '../../../store/system/types';
+import { RootState } from '../../store';
+import { MobileView } from '../../store/system/types';
 
 import LobbyDesktop from './lobbyDesktop';
 import LobbyMobile from './lobbyMobile';
