+++ packages/frontend/components/lobby/mainPanel/chat.tsx
@@ -1,12 +1,12 @@
 import React, { ReactElement, useState, useRef, useEffect } from 'react';
 import { connect } from 'react-redux';
 
-import socket from '../../socket/socket';
+import socket from '../../../socket/socket';
 
-import { RootState } from '../../store';
-import { ThemeOptions } from '../../store/userOptions/types';
+import { RootState } from '../../../store';
+import { ThemeOptions } from '../../../store/userOptions/types';
 import Message from './message';
-import { IMessage } from '../../store/chat/message.types';
+import { IMessage } from '../../../store/chat/message.types';
 
 interface IStateProps {
   theme: ThemeOptions;
