+++ packages/frontend/components/lobby/mainPanel/message.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement } from 'react';
 import { connect } from 'react-redux';
 
-import { RootState } from '../../store';
-import { ThemeOptions } from '../../store/userOptions/types';
-import { IMessage } from '../../store/chat/message.types';
+import { RootState } from '../../../store';
+import { ThemeOptions } from '../../../store/userOptions/types';
+import { IMessage } from '../../../store/chat/message.types';
 
 interface IStateProps {
   theme: ThemeOptions;
