+++ packages/frontend/components/lobby/rightPanel/gameCard.tsx
